frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "python-barcode>=0.15.1",
    "requests>=2.32.3",
    "setuptools>=77.0.1",
    "streamlit>=1.43.2",
]
